Tests/LibWeb/Text/input/Streams/ReadableStream-pipeThrough-cannot-pipe-to-locked-stream.html

<script src="../include.js"></script>
<script>
    asyncTest(done => {
        const transformStream = new TransformStream();

        const stream = new ReadableStream();

        transformStream.writable.getWriter();
        try {
            stream.pipeThrough(transformStream);
        }
        catch (e) {
            println(e);
        }
        done();
    });
</script>
